Change name of test file, removed redundant code in file
