+++ Assets/Tests/BombRangeTest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a5d0e10245313484a8872a99f448cd71
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
